ARTEMIS-2190 move tests
The "jms-tests" module is deprecated and these tests should have never
gone in there. Moving them to the "integration-tests" module.

(cherry picked from commit d7eb076d4bbde61a68617dfcee0f790ef873b05c)
(cherry picked from commit 29abe0d8c17cd5cdf927474f6ef488c47f17a3e2)

downstream: ENTMQBR-2190
